% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CV_criterion}
\alias{CV_criterion}
\title{Cross-Validation Criterion}
\usage{
CV_criterion(
  m,
  m_tilde,
  M,
  Cov,
  Y,
  Delta,
  sigma0,
  tau,
  starting_beta,
  trials,
  nfolds
)
}
\arguments{
\item{m, m_tilde}{Integers which specify the model dimension in the density.}

\item{M}{Integer which specifies the degree of the heteroskedasticity that is
estimated (\code{"M=0"} fits a homoskedastic model)}

\item{Cov}{Matrix with covariates: Every row corresponds to an observation,
each column corresponds to a covariate. No column of Cov is allowed to be
constant, the intercept is added automatically. If no covariates shall be
used, this should be set to NULL.}

\item{Y}{Vector of responses, \code{"length(Y)"} must equal
\code{"nrow(Cov)"}}

\item{Delta}{Vector of censoring indicators, \code{"Delta[i]=1"} means that
observation i is not censored, \code{"length(Delta)"} must equal
\code{"length(Y)"}}

\item{sigma0}{Intercept that is added to the heteroskedasticity function.}

\item{tau}{Quantile of interest, element of (0,1)}

\item{starting_beta}{If provided this is the initial value for beta used for
the optimization routine, if FALSE, a starting value is computed by letting
\code{"m=m_tilde=0"}}

\item{trials}{Number of random starting points for theta, theta_tilde and
lambda in the optimization}

\item{nfolds}{The number of folds to be used in the cross-validation.}
}
\value{
The value of the cross-validation criterion divided by
  \code{"nfolds"}.
}
\description{
`cv_criterion` Computes the cross-validation criterion for given model
dimensions m, m_tilde and M.
}
\details{
Note that the splitting of the data is random and happens in this function.
Therefore consecutive calls of this function will yield different results.
}
\examples{
epsilon <- rnorm(500,mean=0,sd=1)
x     <- runif(500,0-1,1)
T     <- 3+5*x+epsilon
C     <- runif(500,min=0,max=7)
Y     <- pmin(T,C)
Delta <- as.numeric(T==Y)
CV_criterion(2,2,2,x,Y,Delta,0.1,0.5,FALSE,32,5)

}
