% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{laguerre_cross_validation}
\alias{laguerre_cross_validation}
\title{Cross Validation}
\usage{
laguerre_cross_validation(
  Y,
  Delta,
  tau,
  sigma0,
  print.level = 0,
  maxdim = 10,
  Cov = NULL,
  starting_beta = FALSE,
  trials = 32,
  nfolds = 5
)
}
\arguments{
\item{Y}{Vector of responses, \code{"length(Y)"} must equal
\code{"nrow(Cov)"}}

\item{Delta}{Vector of censoring indicators, \code{"Delta[i]=1"} means that
observation i is not censored, \code{"length(Delta)"} must equal
\code{"length(Y)"}}

\item{tau}{Quantile of interest, element of (0,1)}

\item{sigma0}{Intercept that is added to the heteroskedasticity function.}

\item{print.level}{If 0 (the default) no status information is printed, for
all other values the current step is printed.}

\item{maxdim}{Integer dimension at which the optimisation shall stop at the
latest: If either m, m_tilde or M gets larger than maxdim, the optimisation
is stopped and the current values of m, m_tilde and M are returned. The
default is 10.}

\item{Cov}{Matrix with covariates: Every row corresponds to an observation,
each column corresponds to a covariate. No column of \code{"Cov"} is
allowed to be constant, the intercept is added automatically. If no
covariates shall be used, this should be set to NULL (the default).}

\item{starting_beta}{If provided this is the initial value for beta used for
the optimization routine, if FALSE (the default), a starting value is
computed by letting \code{"m=m_tilde=0"}}

\item{trials}{Number of random starting points for theta, theta_tilde and
lambda in the optimization (the default is 32).}

\item{nfolds}{The number of folds to be used in the cross-validation (the
default is 5).}
}
\value{
List of five elements: m, m_tilde and M contain the found degrees and
  est contains the output of \code{"laguerre_estimator"} for these degrees.
  The element path is a matrix od four columns: The first three columns
  contain values of m, m_tilde and M and the fourth column show the
  corresponding value of the cross-validation criterion.
}
\description{
`laguerre_cross_validation` performs Cross-Validation for the dimensions of
the approximation m, m_tilde and M.
}
\details{
The function [laguerre_estimator()] is repeatedly called with different
values for \code{"m"}, \code{"m_tilde"} and \code{"M"}. The other values are
always the same as specified in the call of this function. Note that the data
splitting itself happens in the function [CV_criterion()] as described there.
This functions is just an optimization algorithm for \code{"CV_criterion"}.
}
\examples{
epsilon <- rnorm(500,mean=0,sd=1)
x     <- runif(500,0-1,1)
T     <- 3+5*x+epsilon
C     <- runif(500,min=0,max=7)
Y     <- pmin(T,C)
Delta <- as.numeric(T==Y)
laguerre_cross_validation(Y,Delta,0.5,0.1,1,Cov=x)

}
