% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{laguerre_estimator}
\alias{laguerre_estimator}
\title{Laguerre Estimator}
\usage{
laguerre_estimator(
  m,
  m_tilde,
  M = 0,
  Cov = NULL,
  Y,
  Delta,
  sigma0 = 0.2,
  tau,
  starting_beta = FALSE,
  trials = 32
)
}
\arguments{
\item{m, m_tilde}{Integers which specify the model dimension in the density}

\item{M}{Integer which specifies the degree of the heteroskedasticity that is estimated (\code{"M=0"}, the default, fits a homoskedastic model)}

\item{Cov}{Matrix with covariates: Every row corresponds to an observation, each column corresponds to a covariate. No column of Cov is allowed to be constant, the intercept is added automatically. If no covariates shall be used, this should be set to NULL, the default.}

\item{Y}{Vector of responses, \code{"length(Y)"} must equal \code{"nrow(Cov)"}}

\item{Delta}{Vector of censoring indicators, \code{"Delta[i]=1"} means that observation i is not censored, \code{"length(Delta)"} must equal \code{"length(Y)"}}

\item{sigma0}{Intercept that is added to the heteroskedasticity function, default is 0.2.}

\item{tau}{Quantile of interest, element of (0,1)}

\item{starting_beta}{If provided this is the initial value for beta used for the optimization routine, if FALSE, the default, a starting value is computed by letting \code{"m=m_tilde=0"}}

\item{trials}{Number of random starting points for theta, theta_tilde and lambda in the optimization (default is 32)}
}
\value{
The function returns a list with the five elements "objective" (optimal value of likelihood), "beta" (estimator for beta), "theta", "theta_tilde", "lambda" (estimates for the corresponding parameters).
}
\description{
`laguerre_estimator` computes the Laguerre estimator for given degrees m, m_tilde and M. For a detailed description of how the optimisation works we refer to the documentation pdf.
}
\examples{
epsilon <- rnorm(500,mean=0,sd=1)
x     <- runif(500,0-1,1)
T     <- 3+5*x+epsilon
C     <- runif(500,min=0,max=7)
Y     <- pmin(T,C)
Delta <- as.numeric(T==Y)
laguerre_estimator(m=0,m_tilde=0,M=0,Cov=x,Y=Y,Delta=Delta,sigma0=0.1,tau=0.5,starting_beta=FALSE,trials=32)

}
